update map gen, make music loop, cursor only show if bomb equipped

+++ game_defence/src/modules/system_turnbased_enemy/helpers.cpp
@@ -24,7 +24,7 @@ move_action_common(entt::registry& r, const entt::entity e, const glm::vec2& dst
 
   // move action...
   const auto it = std::find(map.map.begin(), map.map.end(), e);
-  const auto idx = it - map.map.begin();
+  const auto idx = static_cast<int>(it - map.map.begin());
   const int a = static_cast<int>(idx);
   const int b = next_idx;
